test7d: unsigned integer divide instruction
This also changes the -mcpu flag for cpu-arm.S. Given that's effectively just
telling the assembler what instruction sets are available (and not the
compiler) I don't anticipate breakage (except on udiv itself) on A8/A9. Going
to test in a moment anyway, though.

I'm also going to look at handling detection of the udiv instruction.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>